skillvision/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai[aiohttp]>=1.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pypdf>=4.0.0
